requirements.txt

---

## 📄 `requirements.txt`
```text
selenium
python-dotenv
